\name{Disake-package}
\alias{Disake-package}
\alias{Disake}
\docType{package}
\title{ Discrete associated kernel estimators.
}
\description{
Discrete kernel probability mass function estimation is performed through three standard discrete associated kernels: Binomial, Poisson and Negative binomial kernels. The cross validation method for bandwidth selection is also implemented.
}
\details{
\tabular{ll}{
Package: \tab Disake\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-20\cr
License:GPL (>= 2)
}

}
\author{
W. E. Wansouwe , C. C. Kokonendji and D.T. Kolyang 

Maintainer: W. E. Wansouwe <ericwansouwe@gmail.com>

}
\references{

Abdous, B. and Kokonendji, C.C. (2009).  Consistency and asymptotic normality for discrete associated kernel estimator, African
Diaspora Journal of Mathematics 8,  63 - 70.

Aitchison, J. and Aitken, C.G.G. (1976).  Multivariate binary discrimination by the kernel method, Biometrika 63, 413-420.

Kokonendji, C.C. and Senga Kiesse, T. (2011). Discrete associated kernel method and extensions.
Statistical Methodology 8,  497 - 516.

Kokonendji, C.C., Senga Kiesse, T. and Zocchi, S.S. (2007). Discrete triangular distributions and non-parametric estimation for
probability mass function.
Journal of Nonparametric Statistics 19,  241 - 254.


C.C. Kokonendji, S.S. Zocchi, Extensions of discrete triangular distributions and boundary bias in kernel estimation for
discrete functions, Statistics & Probability Letters 80 (2010) 1655 - 1662.

Zougab, N., Adjabi, S. and Kokonendji, C.C. (2012). Binomial kernel and Bayes local bandwidth in discrete functions estimation. 
Journal of Nonparametric Statistics 24, 783 - 795.




}

\keyword{ package }

